Reorganized vacation package
- introduces vacation.NopArchive type for non-archive files
- splits vacation implementation files up into type-named files
- renames vacation test files to remove stutter
- adds vacation.Archive.WithName option to allow NopArchives to have
  specified names
- uses vacation.Archive.WithName option in postal.Service.Deliver to
  deliver file to location matching the dependency URI basename
- includes application/jar mime-type as NopArchive type
