+++ vacation/symlink_sorting_test.go
@@ -15,7 +15,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func testVacationSymlinkSorting(t *testing.T, context spec.G, it spec.S) {
+func testSymlinkSorting(t *testing.T, context spec.G, it spec.S) {
 	var (
 		Expect = NewWithT(t).Expect
 	)
