+++ vacation/tar_gzip_archive_test.go
@@ -3,6 +3,7 @@ package vacation_test
 import (
 	"archive/tar"
 	"bytes"
+	"compress/gzip"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -10,20 +11,19 @@ import (
 
 	"github.com/paketo-buildpacks/packit/vacation"
 	"github.com/sclevine/spec"
-	"github.com/ulikunitz/xz"
 
 	. "github.com/onsi/gomega"
 )
 
-func testVacationTarXZ(t *testing.T, context spec.G, it spec.S) {
+func testTarGzipArchive(t *testing.T, context spec.G, it spec.S) {
 	var (
 		Expect = NewWithT(t).Expect
 	)
 
-	context("TarXZArchive.Decompress", func() {
+	context("Decompress", func() {
 		var (
-			tempDir      string
-			tarXZArchive vacation.TarXZArchive
+			tempDir        string
+			tarGzipArchive vacation.TarGzipArchive
 		)
 
 		it.Before(func() {
@@ -32,10 +32,8 @@ func testVacationTarXZ(t *testing.T, context spec.G, it spec.S) {
 			Expect(err).NotTo(HaveOccurred())
 
 			buffer := bytes.NewBuffer(nil)
-			xzw, err := xz.NewWriter(buffer)
-			Expect(err).NotTo(HaveOccurred())
-
-			tw := tar.NewWriter(xzw)
+			gw := gzip.NewWriter(buffer)
+			tw := tar.NewWriter(gw)
 
 			Expect(tw.WriteHeader(&tar.Header{Name: "some-dir", Mode: 0755, Typeflag: tar.TypeDir})).To(Succeed())
 			_, err = tw.Write(nil)
@@ -61,9 +59,9 @@ func testVacationTarXZ(t *testing.T, context spec.G, it spec.S) {
 			Expect(err).NotTo(HaveOccurred())
 
 			Expect(tw.Close()).To(Succeed())
-			Expect(xzw.Close()).To(Succeed())
+			Expect(gw.Close()).To(Succeed())
 
-			tarXZArchive = vacation.NewTarXZArchive(bytes.NewReader(buffer.Bytes()))
+			tarGzipArchive = vacation.NewTarGzipArchive(bytes.NewReader(buffer.Bytes()))
 
 		})
 
@@ -73,7 +71,7 @@ func testVacationTarXZ(t *testing.T, context spec.G, it spec.S) {
 
 		it("unpackages the archive into the path", func() {
 			var err error
-			err = tarXZArchive.Decompress(tempDir)
+			err = tarGzipArchive.Decompress(tempDir)
 			Expect(err).ToNot(HaveOccurred())
 
 			files, err := filepath.Glob(fmt.Sprintf("%s/*", tempDir))
@@ -100,7 +98,7 @@ func testVacationTarXZ(t *testing.T, context spec.G, it spec.S) {
 
 		it("unpackages the archive into the path but also strips the first component", func() {
 			var err error
-			err = tarXZArchive.StripComponents(1).Decompress(tempDir)
+			err = tarGzipArchive.StripComponents(1).Decompress(tempDir)
 			Expect(err).ToNot(HaveOccurred())
 
 			files, err := filepath.Glob(fmt.Sprintf("%s/*", tempDir))
@@ -111,16 +109,15 @@ func testVacationTarXZ(t *testing.T, context spec.G, it spec.S) {
 
 			Expect(filepath.Join(tempDir, "some-other-dir")).To(BeADirectory())
 			Expect(filepath.Join(tempDir, "some-other-dir", "some-file")).To(BeARegularFile())
-
 		})
 
 		context("failure cases", func() {
-			context("when it fails to create a xz reader", func() {
+			context("when it fails to create a grip reader", func() {
 				it("returns an error", func() {
-					readyArchive := vacation.NewTarXZArchive(bytes.NewBuffer([]byte(`something`)))
+					readyArchive := vacation.NewTarGzipArchive(bytes.NewBuffer([]byte(`something`)))
 
 					err := readyArchive.Decompress(tempDir)
-					Expect(err).To(MatchError(ContainSubstring("failed to create xz reader")))
+					Expect(err).To(MatchError(ContainSubstring("failed to create gzip reader")))
 				})
 			})
 		})
